requirements.txt
google-generativeai
python-dotenv
spire.pdf
langchain 
langchain-google-genai 
langchain-core
langfuse
